siber_topluluk/requirements.txt
flask==2.3.3
flask-login==0.6.3
flask-pymongo==2.3.0
pymongo==4.6.1
python-dotenv==1.0.0
flask-wtf==1.2.1
email-validator==2.1.0
flask-admin==1.6.1
bcrypt==4.1.1
gunicorn==21.2.0
blinker==1.7.0
werkzeug==2.3.7 